+++ src/main/java/com/telegrambot/WildberriesBot/service/reply/ItemDeleteReplyService.java
@@ -30,7 +30,7 @@ public class ItemDeleteReplyService implements Reply {
     @Override
     public SendMessage sendMessage(Message message) {
         long chatId = message.getChatId();
-        int userId = message.getFrom().getId();
+        long userId = message.getFrom().getId();
         try {
             long a = Integer.parseInt(message.getText());
         } catch (RuntimeException e) {
